+++ crates/astria-core/CHANGELOG.md
@@ -61,6 +61,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `TRANSACTION_FAILED_EXECUTION` [#2142](https://github.com/astriaorg/astria/pull/2142).
 - Change `CurrencyPairsChange` action to hold an `IndexSet` rather than a `Vec`
   of currency pairs [#2171](https://github.com/astriaorg/astria/pull/2171).
+- Update `celestia-types` dependency [#2253](https://github.com/astriaorg/astria/pull/2253).
 
 ### Removed