+++ crates/astria-sequencer-relayer/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update `celestia-types` dependency [#2253](https://github.com/astriaorg/astria/pull/2253).
+
 ## [1.0.1] - 2025-03-06
 
 ### Changed